Move FilterItem to a separate package as it is not auto-generated.
