+++ common/core/src/commonMain/kotlin/com/trafi/core/model/ext/FilterItem.kt
@@ -1,4 +1,4 @@
-package com.trafi.core.model
+package com.trafi.core.model.ext
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
